Random things for uikit

+++ src/button.rs
@@ -37,7 +37,7 @@ impl UIButton {
         let cls = Self::class();
         unsafe {
             let obj: *mut Self = msg_send![cls, alloc];
-            let obj: *mut Self = msg_send![obj, buttonWithType:(button_type as i64)];
+            let obj: *mut Self = msg_send![obj, buttonWithType:button_type];
             //let obj: Result<*mut Self, MessageError> = (*obj).send_message(sel!(buttonWithType:), (button_type as i64));
             Id::from_retained_ptr(obj)
         }
